+++ packages/filesystem/src/browser/filesystem-frontend-module.ts
@@ -38,8 +38,10 @@ import { FileSystemWatcherErrorHandler } from './filesystem-watcher-error-handle
 import { UTF8 } from '@theia/core/lib/common/encodings';
 import { FilepathBreadcrumbsContribution } from './breadcrumbs/filepath-breadcrumbs-contribution';
 import { BreadcrumbsFileTreeWidget, createFileTreeBreadcrumbsWidget } from './breadcrumbs/filepath-breadcrumbs-container';
+import { FilesystemSaveResourceService } from './filesystem-save-resource-service';
+import { SaveResourceService } from '@theia/core/lib/browser/save-resource-service';
 
-export default new ContainerModule(bind => {
+export default new ContainerModule((bind, unbind, isBound, rebind) => {
     bindFileSystemPreferences(bind);
 
     bindContributionProvider(bind, FileServiceContribution);
@@ -224,6 +226,9 @@ export default new ContainerModule(bind => {
     );
     bind(FilepathBreadcrumbsContribution).toSelf().inSingletonScope();
     bind(BreadcrumbsContribution).toService(FilepathBreadcrumbsContribution);
+
+    bind(FilesystemSaveResourceService).toSelf().inSingletonScope();
+    rebind(SaveResourceService).toService(FilesystemSaveResourceService);
 });
 
 export function bindFileResource(bind: interfaces.Bind): void {